python-backend/pyproject.toml
[project]
name = "python-backend"
version = "0.1.0"
description = "FastAPI backend for LearnMe AI operations"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.5.0",
    "fastapi>=0.121.1",
    "uvicorn>=0.38.0",
    "python-dotenv>=1.2.1",
    "pydantic-settings>=2.11.0",
    "pydantic>=2.12.4",
    "beautifulsoup4>=4.14.2",
    "httpx>=0.28.1",
]
